requirements.txt
# Core Requirements
fastapi
uvicorn
python-multipart
pydantic

# AI/ML Dependencies
huggingface-hub
langchain-core
langchain-community
langchain-groq
langchain_huggingface
langgraph
python-dotenv
sentence-transformers
pypdf
langchain-chroma
wikipedia
duckduckgo-search
torch
tensorflow
transformers
numpy

# Vector Database
chromadb

# Optional (for production)
python-jose
passlib
gunicorn
flask
pytest